Remove files for pages that are no longer exposed
The URLs for these pages have redirected for a while, or have been
unused or no longer relevant for some time.

I've updated redirects.json so that more of this is testable locally
and in stage rather than relying (only) on overrides at Cloudflare,
which can mostly be removed once this is deployed.

Ref https://github.com/jquery/jquery.org/issues/139
Closes https://github.com/jquery/jquery.org/pull/118